+++ My project/Assets/IInteractable.cs
@@ -1,0 +1,7 @@
+﻿namespace DefaultNamespace
+{
+    public class IInteractable
+    {
+        
+    }
+}
