+++ My project/Assets/IInteractable.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 528cf7b049664d77acbcb1853d97e34a
+timeCreated: 1681125128
