+++ My project/Assets/InteractionDetector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5305b3d054e4a3c428b5d9fa0ae9287f
+guid: ef1eff60ece13e0429cfc2549d004130
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
