--- My project/Assets/h.cs
@@ -1,34 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class h : MonoBehaviour
-{
-    public float speed;
-    // Start is called before the first frame update
-    void Start()
-    {
-        transform.position = new Vector3(0,0,0);
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        // if(transform.position.y>=0){
-        //     transform.position = new Vector3(transform.position.x,0,0);
-        // }
-        // else if(transform.position.y <= -4.3f){
-        //     transform.position = new Vector3(transform.position.x, -4.3f, 0);
-        // }
-
-        // if(transform.position.x>=0){
-        //     transform.position = new Vector3(0,transform.position.y,0);
-        // }
-        // else if(transform.position.x <= -4.3f){
-        //     transform.position = new Vector3(-4.3f, transform.position.y, 0);
-        // }
-
-        transform.Translate(Vector3.right*Time.deltaTime*speed*Input.GetAxis("Horizontal"));
-        transform.Translate(Vector3.up*Time.deltaTime*speed*Input.GetAxis("Vertical"));
-    }
-}
